pyproject.toml
[tool.black]
line-length = 80

[tool.poetry]
name = "flask-template"
version = "0.1.0"
description = "Template to Flask Application"
authors = ["Revton Braga <revtonbr@gmail.com>"]
readme = "README.md"
repository = "https://github.com/revton/flask-template"
homepage = "https://flask-template.com.br"
documentation = "https://flask-template.com.br/docs"
license = "MIT"
include = ["README.md"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
python-dotenv = "^0.17.0"
gunicorn = "^20.1.0"
gevent = "^21.1.2"
dynaconf = "^3.1.4"
Flask-Admin = "^1.5.8"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^2.7.0"
flask-restx = "^0.3.0"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.24.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
isort = "^5.8.0"
flake8 = "^3.9.0"
ipdb = "^0.13.7"
safety = "^1.10.3"
coverage = "^5.5"
Flask-DebugToolbar = "^0.11.0"
radon = "^4.5.0"
flake8-polyfill = "^1.0.2"
behave = "^1.2.6"
selenium = "^3.141.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"